+++ CHANGES.md
@@ -1,5 +1,15 @@
 ## Change Log
 
+### 0.0.2 - 2021-11-23
+
+#### Added
+
+- Initialize **"Dynamic Graph Rendering"** resource
+
+#### Changed
+
+- Update minor content of **"Multi-level Nested Simulation"** resource and **"Force Circular Sector"** resource
+
 ### 0.0.1 - 2021-11-20
 
 #### Added